feat: improved training
